--- src/app/(user)/main/_layout.tsx
@@ -1,11 +1,0 @@
-import { Stack } from 'expo-router';
-
-const MainLayout = () => {
-  return (
-    <Stack>
-      <Stack.Screen name='index' options={{ headerShown: false }} />
-    </Stack>
-  );
-};
-
-export default MainLayout;
