--- src/app/(user)/main/index.tsx
@@ -1,7 +1,0 @@
-import MainScreen from '../../../components/screen/Main/MainScreen';
-
-const MainIndex = () => {
-  return <MainScreen />;
-};
-
-export default MainIndex;
